requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-extensions==3.1.3
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
djlint==0.5.8
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.3.1
pylama==7.7.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.10.8
requests==2.27.1
six==1.16.0
snowballstemmer==2.1.0
sqlparse==0.4.2
toml==0.10.2
tomlkit==0.7.2
tqdm==4.62.3
uritemplate==4.1.1
urllib3==1.26.9
